Buffer writes to SQL database for speed
Queue the writes to the sql database to keep it open less, and reduce number of I/O operations.
P.S.: Hey landmaster, was trying to figure out how to contact you and figured this would be a decent way to get in contact. I had been working on a website similar to btd6index, but more general purpose, when i saw you had made btd6index.win. Not sure if you'd be interested, but I had begun work on a website that includes things such as a megalist of challenge solutions (with hints but that takes a ton of time so its on the back burner), 2tcs, a gui for quincybot's features (and some more, like a hero xp reverse calculator [figure out when a hero needs to be placed to hit a certain level by a certain round]). If you're interested in talking, you can either continue the conversation here, or reach out to me on discord. My username is wordsmith123. If you don't have discord, I'd be happy to provide my email on request. Cheers.